+++ tests/Feature/GroupTest.php
@@ -17,3 +17,25 @@ test("routes with a group", function () {
 
     expect($this->output)->toBe("profile");
 });
+
+
+test("routes with deep group", function () {
+    $id = "123";
+    $this->mock("post", "https://www.example.com/users/profile/photo/save/$id");
+    
+    $router = new Router;
+
+    $router->prefix("/users")->group(function () use($router) {
+        $router->prefix("/profile")->group(function () use($router) {
+            $router->prefix("/photo")->group(function () use($router) {
+                $router->post("/save/{id}", function ($id) {
+                    $this->output = $id;
+                });
+            });
+        });
+    });
+
+    $router->dispatch($this->request());
+
+    expect($this->output)->toBe($id);
+});