+++ tests/TestCase.php
@@ -9,7 +9,7 @@ use PHPUnit\Framework\TestCase as BaseTestCase;
 abstract class TestCase extends BaseTestCase
 {
 
-    public mixed $output;
+    public mixed $output = null;
     
     public function mock(string $method, $url)
     {